requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.8
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-six==1.0.4
djangorestframework==3.12.4
idna==3.2
pytz==2021.3
requests==2.26.0
screen==1.0.1
sqlparse==0.4.2
urllib3==1.26.7
xlwt==1.3.0
